[hotfix][yarn] Add AbstractClusterClientFactory containing some common implementations
