+++ flink-yarn/src/main/java/org/apache/flink/yarn/YarnClusterClientFactory.java
@@ -19,13 +19,10 @@
 package org.apache.flink.yarn;
 
 import org.apache.flink.annotation.Internal;
+import org.apache.flink.client.deployment.AbstractClusterClientFactory;
 import org.apache.flink.client.deployment.ClusterClientFactory;
-import org.apache.flink.client.deployment.ClusterSpecification;
 import org.apache.flink.configuration.Configuration;
-import org.apache.flink.configuration.ConfigurationUtils;
 import org.apache.flink.configuration.DeploymentOptions;
-import org.apache.flink.configuration.TaskManagerOptions;
-import org.apache.flink.runtime.clusterframework.TaskExecutorResourceUtils;
 import org.apache.flink.yarn.configuration.YarnConfigOptions;
 import org.apache.flink.yarn.executors.YarnJobClusterExecutor;
 import org.apache.flink.yarn.executors.YarnSessionClusterExecutor;
@@ -43,7 +40,7 @@ import static org.apache.flink.util.Preconditions.checkNotNull;
  * A {@link ClusterClientFactory} for a YARN cluster.
  */
 @Internal
-public class YarnClusterClientFactory implements ClusterClientFactory<ApplicationId> {
+public class YarnClusterClientFactory extends AbstractClusterClientFactory<ApplicationId> {
 
 	@Override
 	public boolean isCompatibleWith(Configuration configuration) {
@@ -65,28 +62,6 @@ public class YarnClusterClientFactory implements ClusterClientFactory<Applicatio
 		checkNotNull(configuration);
 		final String clusterId = configuration.getString(YarnConfigOptions.APPLICATION_ID);
 		return clusterId != null ? ConverterUtils.toApplicationId(clusterId) : null;
-	}
-
-	@Override
-	public ClusterSpecification getClusterSpecification(Configuration configuration) {
-		checkNotNull(configuration);
-
-		// JobManager Memory
-		final int jobManagerMemoryMB = ConfigurationUtils.getJobManagerHeapMemory(configuration).getMebiBytes();
-
-		// Task Managers memory
-		final int taskManagerMemoryMB = TaskExecutorResourceUtils
-			.resourceSpecFromConfig(configuration)
-			.getTotalProcessMemorySize()
-			.getMebiBytes();
-
-		int slotsPerTaskManager = configuration.getInteger(TaskManagerOptions.NUM_TASK_SLOTS);
-
-		return new ClusterSpecification.ClusterSpecificationBuilder()
-				.setMasterMemoryMB(jobManagerMemoryMB)
-				.setTaskManagerMemoryMB(taskManagerMemoryMB)
-				.setSlotsPerTaskManager(slotsPerTaskManager)
-				.createClusterSpecification();
 	}
 
 	private YarnClusterDescriptor getClusterDescriptor(Configuration configuration) {
